refactor: renamed components and modified imports
