+++ web/src/components/SurveyForm/SurveyDataForm.tsx
@@ -4,8 +4,8 @@ import styled, { DefaultTheme } from 'styled-components';
 
 import { SurveyCreateRequest, CertificationType } from '../../types/request/Survey';
 import { NumberUtils } from '../../utils/NumberUtils';
+import { PlusImage } from '../Button/ImageButtons';
 import CertificationList from '../CertificationList';
-import { PlusImage } from '../Styled/ImageButtons';
 
 const Container = styled.div`
   display: flex;
